Fix a bug relating to activeplayer
In the previous version, active player would switch even if the player
clicked on a spot that wasn't free. The fix is more of a workaround,
as now checking for winning state and player switching will be done
inside of the addMark function within the playerFactory function.
Unless a better alternative is discovered.